% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{calc_deviation}
\alias{calc_deviation}
\title{Calculate the deviation from the mean}
\usage{
calc_deviation(data = NULL, meta_data = NULL)
}
\arguments{
\item{data}{data.frame, with all data.}

\item{meta_data}{character() vector with column names of the meta data.}
}
\value{
data.frame in long format
}
\description{
Calculate the deviation from the mean for each QC sample overall
   and per batch.
}
\author{
Rico Derks
}
