% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{calc_trend}
\alias{calc_trend}
\title{Do the calculations for a trend plot}
\usage{
calc_trend(data = NULL, meta_data = NULL)
}
\arguments{
\item{data}{data.frame, with all data.}

\item{meta_data}{character() vector with column names of the meta data.}
}
\value{
data.frame in long format with log2fc column
}
\description{
Do the calculations for a trend plot
}
\author{
Rico Derks
}
