% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{deviation_plot}
\alias{deviation_plot}
\title{Create a deviation from the mean plot}
\usage{
deviation_plot(data = NULL, trend = c("overall", "batch"))
}
\arguments{
\item{data}{data.frame, with all data.}

\item{trend}{character(1) show the overall trend or per batch.}
}
\value{
ggplot2 object
}
\description{
Create a deviation from the mean histogram, overall or per batch.
}
\author{
Rico Derks
}
