% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{trend_plot}
\alias{trend_plot}
\title{Do the calculations for a trend plot}
\usage{
trend_plot(data = NULL, trend = c("overall", "batch"))
}
\arguments{
\item{data}{data.frame, with all data.}

\item{trend}{character(1) show the overall trend or per batch.}
}
\value{
ggplot2 object
}
\description{
Do the calculations for a trend plot
}
\author{
Rico Derks
}
